+++ src/UpForm/UpSchemaObject.tsx
@@ -43,7 +43,6 @@ export interface UpSchemaObjectState {
   showAdvanced: boolean;
 }
 
-
 const getStyles = (props) => {
   const { columnSpacing, columnNumber, rowSpacing } = props;
   return style({
@@ -69,14 +68,23 @@ function compareItems<T extends { order: number }>(a: T, b: T): number {
   return -1;
 }
 
-export function groupByRow<
-  T extends { colspan?: number; order: number; isSeparator?: boolean }
->(items: T[], defaultColspan: number): T[][] {
-  let usedColSpan = 0;
-  let colspan = 0;
-  const spanLimit = 24;
-  const rows = items.sort(compareItems).reduce((rows: T[][], viewModel: T) => {
-    colspan = viewModel.colspan || defaultColspan ;
+export function manageColspan<
+  T extends {
+    colspan?: number;
+    order: number;
+    isSeparator?: boolean;
+    group?: string;
+  }
+>(items: T[], defaultColspan: number, group?: string) {
+  return items.sort(compareItems).reduce((rows, viewModel) => {
+    let usedColSpan = 0;
+    let colspan = 0;
+    const spanLimit = 24;
+    if (group) {
+      viewModel.group = group === "undefined" ? "Autre" : group;
+    }
+
+    colspan = viewModel.colspan || defaultColspan;
     usedColSpan += colspan;
     let currentRow;
     if (rows.length === 0) {
@@ -85,11 +93,11 @@ export function groupByRow<
     } else {
       currentRow = rows[rows.length - 1];
     }
-    if(viewModel.isSeparator) {
-      usedColSpan = spanLimit
+    if (viewModel.isSeparator) {
+      usedColSpan = spanLimit;
     }
 
-    if (usedColSpan > spanLimit ) {
+    if (usedColSpan > spanLimit) {
       currentRow = [];
       rows.push(currentRow);
       usedColSpan = colspan;
@@ -98,10 +106,32 @@ export function groupByRow<
     if (!viewModel.isSeparator) {
       currentRow.push(viewModel);
     }
-    
+
     return rows;
   }, []);
-  return rows;
+}
+
+export function groupByRow<
+  T extends {
+    colspan?: number;
+    order: number;
+    isSeparator?: boolean;
+    group?: string;
+  }
+>(items: T[], defaultColspan: number): T[][] {
+  const oneGroupAtLeast = items.find((item) => !_.isUndefined(item.group));
+
+  if (!_.isUndefined(oneGroupAtLeast)) {
+    const groupedByGroup = _.groupBy(items, "group");
+
+    const managedGroup = _.forIn(groupedByGroup, (values, key) => {
+      manageColspan(values, defaultColspan, key);
+    });
+    return _.values(managedGroup);
+  } else {
+    const rows = manageColspan(items, defaultColspan);
+    return rows;
+  }
 }
 
 export default class UpSchemaObject extends React.Component<
@@ -124,10 +154,13 @@ export default class UpSchemaObject extends React.Component<
 
   render() {
     let viewModels = (!_.isEmpty(this.props.viewModels) ? this.props.viewModels : this.props.schema['viewModels']) || [] ;
-    
-    const inferViewModels : {name: string, colspan : number, order : number}[] = viewModels.filter(
+    Object.keys(this.props.schema.properties).forEach((key, index) => 
+      this.props.schema.properties[key]['order'] =  this.props.schema.properties[key]['order'] || index+1
+    );
+
+    let inferViewModels : {name: string, colspan : number, order : number,group: string}[] = viewModels.filter(
       a => !this.isIgnored(a.name) && this.props.schema.properties[a.name] && !this.props.schema.properties[a.name].hide
-    ).map( vm => ({...vm, colspan : vm.colspan || this.props.defaultColspan})) ;
+    ).map( vm => ({...vm, order: this.props.schema.properties[vm.name]['order'], group: this.props.schema.properties[vm.name]['group'], colspan :vm.colspan || this.props.defaultColspan})) ;
     
     Object.keys(this.props.schema.properties)
       .filter(
@@ -139,22 +172,41 @@ export default class UpSchemaObject extends React.Component<
             {
               colspan: this.props.defaultColspan,
               name: a,
-              order: inferViewModels.length
+              order: inferViewModels.length,
+              group: this.props.schema.properties[a]['group']
             }
           );
         }
     });
+    
+    inferViewModels = inferViewModels.sort(compareItems);
 
     const rows = groupByRow(
       inferViewModels,
       this.props.defaultColspan
     );
 
+    const groupedRow = rows.map(row =>_.groupBy(row,'group'))
+    
+    let unknownsGroupIndex = -1;
+    groupedRow.forEach((element,index)=>{
+      const unknownGroupExist = element['Autre']
+      if(unknownGroupExist) unknownsGroupIndex = index
+    })
+
+    if(unknownsGroupIndex !== -1) {
+      const unknownGroup = groupedRow[unknownsGroupIndex]
+      groupedRow.splice(unknownsGroupIndex,1)
+      groupedRow.push(unknownGroup)
+    }
+
     let elements = {};
     let elementsAdvanced = [];
+ 
     for (let propertyName in this.props.schema.properties) {
       if (this.props.schema.properties.hasOwnProperty(propertyName)) {
         let property = this.props.schema.properties[propertyName];
+
         if (this.isIgnored(propertyName) || property.hide) continue;
 
         let value =
@@ -199,30 +251,36 @@ export default class UpSchemaObject extends React.Component<
       <UpFormContextConsumer>
         {({ gutter: columnSpacing, rowSpacing }) => (
           <UpGrid gutter={columnSpacing}>
-            {rows.map((row, i) => {
-              return (
-                <UpRow key={i} style={{marginBottom: `${rowSpacing || 10}px`}}>
-                  {this.props.withHR ? <hr /> : null}
-                  {this.props.schema.title == null || this.props.node === "" ? (
-                    ""
-                  ) : (
-                    <h4>{this.props.schema.title}</h4>
-                  )}
-                  {row.map((p, index) => {
-                    return (
-                      <UpCol
-                        key={index}
-                        xs={24}
-                        sm={p.colspan > 12 ? p.colspan : 12}
-                        md={p.colspan}
-                        lg={p.colspan}
-                      >
-                        {elements[p.name]}
-                      </UpCol>
-                    );
-                  })}
-                </UpRow>
-              );
+            {groupedRow.map((group) => {
+              for (let element in group) {
+                const Rows = element !== "undefined" ? (
+                  <fieldset key={element}>
+                    <legend>{element}</legend>
+                    <SchemaRow 
+                      title={this.props.schema.title == null || this.props.node === "" ? "" : this.props.schema.title}
+                      withHR={this.props.withHR}
+                      key={element}
+                      rowSpacing={rowSpacing}
+                      elements={group[element].map(element => ({
+                        colspan: element.colspan,
+                        name: element.name,
+                        render : elements[element.name]
+                      }))} />
+                  </fieldset>
+                ) : (
+                  <SchemaRow 
+                      title={this.props.schema.title == null || this.props.node === "" ? "" : this.props.schema.title}
+                      withHR={this.props.withHR}
+                      key={element}
+                      rowSpacing={rowSpacing}
+                      elements={group[element].map(element => ({
+                        colspan: element.colspan,
+                        name: element.name,
+                        render : elements[element.name]
+                      }))} />
+                );
+                return Rows
+              }
             })}
             {elementsAdvanced != null && elementsAdvanced.length != 0 ? (
               <UpRow>
@@ -283,4 +341,35 @@ export default class UpSchemaObject extends React.Component<
     }
     return required;
   }
+}
+
+interface SchemaRowProps {
+  rowSpacing: number;
+  withHR: boolean;
+  title: string;
+  elements: {colspan : number, name : string, render : React.ReactNode}[];
+}
+
+const SchemaRow : React.FunctionComponent<SchemaRowProps> = ({rowSpacing, withHR, title, elements})  => {
+  return <UpRow style={{ marginBottom: `${rowSpacing || 10}px` }}>
+          {withHR ? <hr /> : null}
+          {title == null ? (
+            ""
+          ) : (
+            <h4>{title}</h4>
+          )}
+          {elements.map((element, index) => {
+            return (
+              <UpCol
+                key={index}
+                xs={24}
+                sm={element.colspan > 12 ? element.colspan : 12}
+                md={element.colspan}
+                lg={element.colspan}
+              >
+                {element.render}
+              </UpCol>
+            );
+          })}
+        </UpRow>
 }